+++ invoker/redis.go
@@ -1,9 +1,0 @@
-package invoker
-
-import (
-	"github.com/integration-system/isp-lib/redis"
-)
-
-var (
-	RedisClient = redis.NewRxClient()
-)