+++ log_code/code.go
@@ -2,4 +2,5 @@ package log_code
 
 const (
 	ErrorDatabaseClient = 801
+	ErrorRedisClient    = 802
 )