+++ redis/client.go
@@ -1,0 +1,33 @@
+package redis
+
+import (
+	rd "github.com/integration-system/isp-lib/redis"
+	"github.com/integration-system/isp-lib/structure"
+	log "github.com/integration-system/isp-log"
+	"isp-system-service/log_code"
+)
+
+var Client = &redisClient{
+	cli: rd.NewRxClient(
+		rd.WithInitHandler(func(c *rd.Client, err error) {
+			if err != nil {
+				log.Fatal(log_code.ErrorRedisClient, err)
+			}
+		})),
+}
+
+type redisClient struct {
+	cli *rd.RxClient
+}
+
+func (c *redisClient) ReceiveConfiguration(configuration structure.RedisConfiguration) {
+	c.cli.ReceiveConfiguration(configuration)
+}
+
+func (c *redisClient) Get() *rd.RxClient {
+	return c.cli
+}
+
+func (c *redisClient) Close() {
+	_ = c.cli.Close()
+}