+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### v2.0.0
-* update to new isp-lib & config service
+### v1.2.0
+* add `access list` methods
+* fix revoke token
 ### v1.1.3
 * update to new log
 * migrate to new db client